Added setMany

+++ contracts/Storage.sol
@@ -18,7 +18,7 @@ contract Storage is Owned {
 
     struct Rate {
         address ratedBy;
-        uint rate;
+        int8 rate;
     }
 
     mapping(bytes32 => Crate) crates;
@@ -77,10 +77,10 @@ contract Storage is Owned {
     }
 
     function setRates(bytes32 _crate, bytes32 _key, address _value, uint _value2) onlyAllowed(_crate) {
-        crates[_crate].rates[_key] = Rate(_value, _value2);
+        crates[_crate].rates[_key] = Rate(_value, int8(_value2));
     }
 
     function getRates(bytes32 _crate, bytes32 _key) constant returns(address, uint) {
-        return (crates[_crate].rates[_key].ratedBy, crates[_crate].rates[_key].rate);
+        return (crates[_crate].rates[_key].ratedBy, uint(crates[_crate].rates[_key].rate));
     }
 }
